backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
pydantic-settings==2.0.3
alembic==1.13.1